+++ pcmdpy/galaxy/galaxy.py
@@ -78,7 +78,7 @@ class CustomGalaxy:
     def set_params(self, gal_params):
         # make sure is array, with right length
         assert(len(gal_params) == self.p_total)
-        gal_params = np.array(gal_params)
+        gal_params = np.array(gal_params, dtype=float)
         # set metal parameters
         feh_params = gal_params[:self.p_feh]
         self.metal_model.set_params(feh_params)
